move input validator to its own package
